move photos data processing to dataIn.js
