+++ Collections/Photos/photos.js
@@ -15,10 +15,11 @@ var logger;
 var fs = require('fs');
 var sync = require('./sync');
 var dataStore = require("./dataStore");
+var dataIn = require("./dataIn");
 
 var lockerInfo;
-var express = require('express'),
-    connect = require('connect');
+var express = require('express');
+var connect = require('connect');
 var app = express.createServer(connect.bodyParser());
 var request = require('request');
 
@@ -118,7 +119,7 @@ app.post('/events', function(req, res) {
         return;
     }
 
-    dataStore.addEvent(req.body, function(err, eventObj) {
+    dataIn.addEvent(req.body, function(err, eventObj) {
         if (err) {
             logger.error("Error processing: " + err);
             res.writeHead(500);
@@ -154,7 +155,7 @@ process.stdin.on('data', function(data) {
     lconfig.load('../../Config/config.json');
     logger = require("logger.js");
     locker.connectToMongo(function(mongo) {
-        sync.init(lockerInfo.lockerUrl, mongo.collections.photo, mongo, locker, lconfig);
+        sync.init(lockerInfo.lockerUrl, mongo, locker, lconfig);
         app.listen(0, function() {
             var returnedInfo = {port: app.address().port};
             process.stdout.write(JSON.stringify(returnedInfo));
